+++ src/pupil_labs/realtime_api/streaming/base.py
@@ -33,6 +33,9 @@ async def receive_raw_rtsp_data(url, *args, **kwargs) -> T.AsyncIterator[RTSPDat
 
 class RTSPRawStreamer:
     def __init__(self, *args, **kwargs):
+        """Forwards all arguments to
+        `aiortsp.rtsp.reader.RTSPReader <https://github.com/marss/aiortsp/blob/master/aiortsp/rtsp/reader.py#L31-L32>`_
+        """
         self._reader = _WallclockRTSPReader(*args, **kwargs)
         self._encoding = None